api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
redis==5.0.8
python-dotenv==1.0.1
